add useShouldLinkToUser hook
